build: migrate from rollup 2 to rolldown
Replaces the rollup 2 + rollup-plugin-typescript2 toolchain with
rolldown 1.2 in both connectkit and connectkit-next-siwe. JS bundling
drops from seconds to milliseconds (~27ms for the 7-entry connectkit
build).

- rolldown.config.{prod,dev}.js replace the rollup configs; same
  entries, output layout (build/*.es.js + chunks/) and sourcemap
  settings, so the package exports/typesVersions maps are unchanged
- externals: all bare specifiers are external via a function, replacing
  rollup-plugin-peer-deps-external plus rollup's unresolved-import
  warnings
- declarations move from rpt2 to `tsc --emitDeclarationOnly` (build
  keeps full type-checking; dev emits types once, then watches JS)
- tsconfig target ES2019 -> ES2020, lib -> es2022 (the tsc step now
  type-checks viem's `ox` dependency, which requires ES2020+)
- siwe/index.ts: `export type` modifiers for type-only re-exports
  (rolldown transforms per-file, isolatedModules semantics)
- drop rollup toolchain from root devDependencies (incl. tslib and the
  dev-only styled-components displayName transformer, which has no oxc
  equivalent)

Verified: build output structure and main-entry import graph identical
to the rollup build (SDK connectors still isolated to their own
entries); vite/nextjs/nextjs-app/testbench examples build; packed
tarball smoke tests pass (SDK-less negative + WC-subpath positive).
