+++ package.json
@@ -27,16 +27,9 @@
     "lint:ci": "yarn workspaces foreach -R run lint"
   },
   "devDependencies": {
-    "@rollup/plugin-node-resolve": "^13.1.3",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
-    "rollup": "^2.80.0",
-    "rollup-plugin-peer-deps-external": "^2.2.4",
-    "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-typescript2": "^0.34.0",
-    "rollup-plugin-visualizer": "^5.5.4",
-    "tslib": "^1.9.3",
-    "typescript-plugin-styled-components": "^2.0.0",
+    "rolldown": "^1.2.0",
     "viem": "^2.55.8",
     "wagmi": "^3.7.4"
   },
