--- packages/connectkit-next-siwe/rollup.config.dev.js
@@ -1,17 +1,0 @@
-import peerDepsExternal from 'rollup-plugin-peer-deps-external';
-import typescript from 'rollup-plugin-typescript2';
-
-import packageJson from './package.json';
-
-export default [
-  {
-    input: ['./src/index.ts'],
-    external: ['react', 'react-dom', 'framer-motion', 'wagmi'],
-    output: {
-      file: packageJson.exports,
-      format: 'esm',
-      sourcemap: false,
-    },
-    plugins: [peerDepsExternal(), typescript()],
-  },
-];
