+++ packages/connectkit-next-siwe/tsconfig.json
@@ -8,8 +8,8 @@
     "declarationDir": "build",
     "module": "ESNext",
     "moduleResolution": "node",
-    "target": "ES2019",
-    "lib": ["es6", "dom", "es2016", "es2017", "es2021"],
+    "target": "ES2020",
+    "lib": ["es2022", "dom"],
     "sourceMap": false,
     "allowSyntheticDefaultImports": true,
     "esModuleInterop": true,
