+++ packages/connectkit/src/siwe/index.ts
@@ -4,4 +4,5 @@ export {
   SIWE_SESSION_QUERY_KEY,
   SIWE_NONCE_QUERY_KEY,
 } from './SIWEProvider';
-export { SIWEContext, SIWEConfig, SIWESession } from './SIWEContext';
+export { SIWEContext } from './SIWEContext';
+export type { SIWEConfig, SIWESession } from './SIWEContext';
